utils: move format_text to textutils
NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring
